backend/requirements.txt
fastapi
uvicorn
gitpython
scikit-learn
google-genai
python-dotenv
